API実装
- UseCase、Controllerなどを実装
- 実装時の設計変更をswaggerに反映

+++ database/migrations/2024_06_22_150129_create_tasks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->text('title');
             $table->string('status');
-            $table->smallInteger('order');
+            $table->unsignedInteger('order');
             $table->timestamps();
             $table->softDeletes();
         });
